Change setTotalPrice

+++ src/java/entity/Order.java
@@ -57,7 +57,7 @@ public class Order {
         return totalPrice;
     }
 
-    public void setTotalPrice(float totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
